Added - Entitybase - derived classes from Kuldemeny

+++ sztf_II_FF/Kuldemenyek/IKuldemeny.cs
@@ -5,6 +5,9 @@ namespace sztf_II_FF.Kuldemenyek
     public interface IKuldemeny : IComparable
     {
         int Prioritas { get; set; }
-        float Tomeg { get; set; }
+        /// <summary>
+        /// Tömeg grammban megadva
+        /// </summary>
+        int Tomeg { get; set; }
     }
 }
